refactor(network): move logic to trackerSummaryUtils
